Multi-Agent-AI/requirements.txt
PyMuPDF==1.26.0
openai>=1.0.0
google-generativeai>=0.3.2
python-dotenv>=1.0.1
